requirements.txt
yt-dlp
google-generativeai
python-dotenv
ffmpeg-python